requirements.txt
playwright
pytest
pytest-html
python-dotenv
eyes-playwright
